OboeTester: add more parameters to auto-latency test
Add channels, sample rate, perf mode, sharing mode.

Fixes #1071
