+++ apps/OboeTester/app/src/main/java/com/google/sample/oboe/manualtest/AnalyzerActivity.java
@@ -45,6 +45,24 @@ public class AnalyzerActivity extends TestInputActivity {
 
     private static final int MY_PERMISSIONS_REQUEST_EXTERNAL_STORAGE = 1001;
 
+    public static final String KEY_IN_SHARING = "in_sharing";
+    public static final String KEY_OUT_SHARING = "out_sharing";
+    public static final String VALUE_SHARING_EXCLUSIVE = "exclusive";
+    public static final String VALUE_SHARING_SHARED = "shared";
+
+    public static final String KEY_IN_PERF = "in_perf";
+    public static final String KEY_OUT_PERF = "out_perf";
+    public static final String VALUE_PERF_LOW_LATENCY = "lowlat";
+    public static final String VALUE_PERF_POWERSAVE = "powersave";
+    public static final String VALUE_PERF_NONE = "none";
+
+    public static final String KEY_IN_CHANNELS = "in_channels";
+    public static final String KEY_OUT_CHANNELS = "out_channels";
+    public static final int VALUE_DEFAULT_CHANNELS = 2;
+
+    public static final String KEY_SAMPLE_RATE = "sample_rate";
+    public static final int VALUE_DEFAULT_SAMPLE_RATE = 48000;
+
     protected static final String KEY_FILE_NAME = "file";
     protected static final String KEY_BUFFER_BURSTS = "buffer_bursts";
 
@@ -179,10 +197,31 @@ public class AnalyzerActivity extends TestInputActivity {
         }
     }
 
-    void configureStreamsFromBundleForApi(Bundle bundle) {
+    private int getPerfFromText(String text) {
+        if (VALUE_PERF_NONE.equals(text)) {
+            return StreamConfiguration.PERFORMANCE_MODE_NONE;
+        } else if (VALUE_PERF_POWERSAVE.equals(text)) {
+            return StreamConfiguration.PERFORMANCE_MODE_POWER_SAVING;
+        } else {
+            return StreamConfiguration.PERFORMANCE_MODE_LOW_LATENCY;
+        }
+    }
+
+    private int getSharingFromText(String text) {
+        if (VALUE_SHARING_SHARED.equals(text)) {
+            return StreamConfiguration.SHARING_MODE_SHARED;
+        } else {
+            return StreamConfiguration.SHARING_MODE_EXCLUSIVE;
+        }
+    }
+
+    void configureStreamsFromBundle(Bundle bundle) {
         // Configure settings
         StreamConfiguration requestedInConfig = mAudioInputTester.requestedConfiguration;
         StreamConfiguration requestedOutConfig = mAudioOutTester.requestedConfiguration;
+
+        requestedInConfig.reset();
+        requestedOutConfig.reset();
 
         // OpenSL ES or AAudio API
         String text = bundle.getString(KEY_IN_API, VALUE_UNSPECIFIED);
@@ -191,6 +230,31 @@ public class AnalyzerActivity extends TestInputActivity {
         text = bundle.getString(KEY_OUT_API, VALUE_UNSPECIFIED);
         audioApi = getApiFromText(text);
         requestedOutConfig.setNativeApi(audioApi);
+
+        // channnels
+        int inChannels = bundle.getInt(KEY_IN_CHANNELS, VALUE_DEFAULT_CHANNELS);
+        requestedInConfig.setChannelCount(inChannels);
+        int outChannels = bundle.getInt(KEY_OUT_CHANNELS, VALUE_DEFAULT_CHANNELS);
+        requestedOutConfig.setChannelCount(outChannels);
+
+        // performance mode
+        text = bundle.getString(KEY_IN_PERF, VALUE_PERF_LOW_LATENCY);
+        int perfMode = getPerfFromText(text);
+        requestedInConfig.setPerformanceMode(perfMode);
+        text = bundle.getString(KEY_OUT_PERF, VALUE_PERF_LOW_LATENCY);
+        perfMode = getPerfFromText(text);
+        requestedOutConfig.setPerformanceMode(perfMode);
+
+        int sampleRate = bundle.getInt(KEY_SAMPLE_RATE, VALUE_DEFAULT_SAMPLE_RATE);
+        requestedInConfig.setSampleRate(sampleRate);
+        requestedOutConfig.setSampleRate(sampleRate);
+
+        text = bundle.getString(KEY_IN_SHARING, VALUE_SHARING_EXCLUSIVE);
+        int sharingMode = getSharingFromText(text);
+        requestedInConfig.setSharingMode(sharingMode);
+        text = bundle.getString(KEY_OUT_SHARING, VALUE_SHARING_EXCLUSIVE);
+        sharingMode = getSharingFromText(text);
+        requestedOutConfig.setSharingMode(sharingMode);
     }
 
     void writeTestResultIfPermitted(String resultString) {
